restore substrate-consensus-aura to always std and address grumbles
